move url builder inside crate
